git-svn-id: https://gentlenav.googlecode.com/svn/branches/MatrixPilot_fbw@2323 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8

+++ MatrixPilot/parameter_table.c
@@ -34,14 +34,14 @@ extern int16_t nav_rollPositionMax ;
 extern int16_t nav_roll_rate ;
 extern uint16_t loiter_radius ;
 extern fractional auto_nav_roll_gain ;
-extern int32_t rate_error_load_gain ;
-extern int32_t pitch_error_rate_gain ;
-extern int32_t yaw_damping ;
-extern int32_t yaw_ff_correction_rate ;
-extern int32_t yaw_ff_correct_thresh ;
-extern int32_t yaw_ffgain_Q16 ;
-extern int32_t roll_damping ;
-extern int32_t roll_error_rate_gain ;
+extern uint32_t rate_error_load_gain ;
+extern uint32_t pitch_error_rate_gain ;
+extern uint32_t yaw_damping ;
+extern uint32_t yaw_ff_correction_rate ;
+extern uint32_t yaw_ff_correct_thresh ;
+extern uint32_t yaw_ffgain_Q16 ;
+extern uint32_t roll_damping ;
+extern uint32_t roll_error_rate_gain ;
 
 
 const mavlink_parameter_parser    mavlink_parameter_parsers[] = {
@@ -111,12 +111,8 @@ const mavlink_parameter mavlink_parameters_list[] = {
     {"FBW_MAX_R_ANGLE" , {.param_int32=20} , {.param_int32=180.0} , UDB_TYPE_INT, PARAMETER_READWRITE, (void*) &fbw_rollPositionMax, sizeof(fbw_rollPositionMax) },
 
     {"NAV_MAX_R_ANGLE" , {.param_int32=20} , {.param_int32=180.0} , UDB_TYPE_INT_CIRCULAR, PARAMETER_READWRITE, (void*) &nav_rollPositionMax, sizeof(nav_rollPositionMax) },
-    {"NAV_ROLL_RATE" , {.param_int32=1} , {.param_int32=180.0} , UDB_TYPE_DCM_ANGLE, PARAMETER_READWRITE, (void*) &
-nav_roll_rate, sizeof(
-nav_roll_rate) },
-    {"NAV_LOITER_RAD" , {.param_int32=0} , {.param_int32=10000.0} , UDB_TYPE_INT, PARAMETER_READWRITE, (void*) &
-loiter_radius, sizeof(
-loiter_radius) },
+    {"NAV_ROLL_RATE" , {.param_int32=1} , {.param_int32=180.0} , UDB_TYPE_DCM_ANGLE, PARAMETER_READWRITE, (void*) &nav_roll_rate, sizeof(nav_roll_rate) },
+    {"NAV_LOITER_RAD" , {.param_int32=0} , {.param_int32=10000.0} , UDB_TYPE_INT, PARAMETER_READWRITE, (void*) &loiter_radius, sizeof(loiter_radius) },
     {"NAV_ROLL_GAIN" , {.param_float=0.01} , {.param_float=1.0} , UDB_TYPE_Q14, PARAMETER_READWRITE, (void*) &auto_nav_roll_gain, sizeof(auto_nav_roll_gain) },
 
     {"GAIN_RATE_LOAD" , {.param_float=-100.0} , {.param_float=100.0} , UDB_TYPE_Q16, PARAMETER_READWRITE, (void*) &rate_error_load_gain, sizeof(rate_error_load_gain) },
@@ -130,7 +126,7 @@ loiter_radius) },
 
     };
 
-const int16_t count_of_parameters_list = sizeof(mavlink_parameters_list) / sizeof(mavlink_parameter);
+const int count_of_parameters_list = sizeof(mavlink_parameters_list) / sizeof(mavlink_parameter);
 
 
 #endif
