+++ libDCM/libDCM_defines.h
@@ -29,7 +29,7 @@ struct absolute2D { int32_t x ; int32_t y ; } ;
 struct relative3D { int16_t x ; int16_t y ; int16_t z ; } ;
 struct absolute3D { int32_t x ; int32_t y ; int32_t z ; } ;
 struct waypoint3D { int32_t x ; int32_t y ; int16_t z ; } ;
-struct fixedOrigin3D {long x; int32_t y; float z;} ;
+struct fixedOrigin3D {int32_t x; int32_t y; float z;} ;
 
 struct dcm_flag_bits {
 			uint16_t unused					: 4 ;
